fix placeholder not following video fit option
It seems as though allowing the parent Stack to be passthrough
forces the placeholder to fit within the bounds of the video's
fit options.

I tested this in my personal app as well as in the example repo
and can confirm that it does not affect other aspects, but would
probably require more rigorous testing.