Bug fix for https://github.com/jhomlala/betterplayer/issues/989
AVURLAsset is particular about network resources without a file extension (e.g. https://fake.dom/HASH). Normally if there is no extension it falls back on the Content-Type. However if a AVAssetResourceLoaderDelegate is used with the AVURLAsset it fails to see the Content-Type of the resource. I suspect that since the delegate manages the URL response, the AVURLAsset never gets the Content-Type and fails. There should be a way to pass this information to the AVURLAsset using the delegate, but for the time being the cleanest working solution is just including the ext in the phony URL already given to the AVURLAsset (which is available due to how BetterPlayer's caching works).