Bug fix for: [BUG] BetterPlayer does not transform correctly in a Transform widget. #998 [BUG] Video texture issue on iOS #965
UiKitView still appears to have issues with Transform widgets. This resolves the issue by rolling back the switch to UiKitView that happened in commit 200777.

As best as I am aware and tested the change was not functionally necessary. The change was not explained as far as I can tell though so possibly there are issues I am not aware of. I have left the UiKitView variant in, commented out in dart, and macro guarded in obj c.